pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accounting-agent-layer"
version = "0.1.0"
description = "Accounting Agent Layer for ERPX (auxiliary accounting workflows)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "Accounting Agent Layer Contributors" }]

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "celery>=5.3",
  "redis>=5.0",
  "httpx>=0.27",
  "tenacity>=8.2",
  "structlog>=24.1",
  "prometheus-client>=0.20",
  "boto3>=1.34",
  "openpyxl>=3.1",
  "rapidfuzz>=3.6",
  "pdfplumber>=0.11",
  "pillow>=10.2",
  "pytesseract>=0.3.10",
  "pdf2image>=1.17",
  "watchdog>=4.0",
  "aiosmtplib>=3.0",
  "python-multipart>=0.0.9",
  "python-dateutil>=2.9",
  "jinja2>=3.1",
  "croniter>=2.0",
  "pyyaml>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
ui = [
  # Static SPA - no Python dependencies needed
]
graphs = [
  "langgraph>=1.0",
  "langchain-core>=0.3",
]
ray = [
  "ray[default]>=2.0,<3.0",
]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.3",
  "mypy>=1.9",
  "types-python-dateutil",
  "types-redis",
  "types-requests",
  "reportlab>=4.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
# FastAPI dependency injection uses `Depends(...)` in default values by design.
"src/accounting_agent/agent_service/main.py" = ["B008"]
"src/accounting_agent/erpx_mock/main.py" = ["B008"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
